Web socket wrapper

+++ GameServer_8/gameserver/FromPlayerMessage.go
@@ -1,6 +1,6 @@
 package gameserver
 
 type FromPlayerMessage struct {
-	ID uint32 `json:"id"`
-	Y  int16  `json:"y"`
+	ID uint32  `json:"id"`
+	Y  float32 `json:"y"`
 }
